% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsca.R
\name{getMeasures}
\alias{getMeasures}
\title{Calculate syntactic measures}
\usage{
getMeasures(input, colnames = c(TOKEN = "TOKEN", POSITION = "POSITION", DEP_ON =
 "DEP_ON", DEP_TYPE = "DEP_TYPE", POS = "POS.MELT", LEMMA = "LEMMA"), round.to = 2)
}
\arguments{
\item{input}{a data frame which contains a sentence in CONLL format or a list
of data frames in CONLL format (see \code{\link{getUnits}})}

\item{colnames}{a named vector containing the names of the following columns:
\itemize{
\item TOKEN: name of the column with the tokens
\item POSITION: name of column with the word index
\item DEP_ON: name of column with the dependency heads
\item DEP_TYPE: name of column with the dependency relations
\item POS: name of column with the part of speech tags
\item LEMMA: name of column with the lemmas
}}

\item{round.to}{integer for rounding measures}
}
\value{
a list of syntactic complexity measures:
\itemize{
\item MLS: mean length of sentence
\item DIVS: standard deviation of sentence length
\item T_S: t-units per sentence
\item MLT: mean length of t-unit
\item DIVT: standard deviation of t-unit length
\item C_T: clauses per sentence
\item MLC: mean length of clause
\item DIVC: standard deviation of clause length
\item MLNP: mean length of noun phrases
\item DIVNP: standard deviation of noun phrase length
\item NP_C: noun prases per clause
}
}
\description{
Calculate syntactic measures
}
\examples{
data(example)
getMeasures(example, round.to = 2)
}
