% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsca.R
\name{getParse}
\alias{getParse}
\title{Extract data frames from output of Malt Parser}
\usage{
getParse(file)
}
\arguments{
\item{file}{a text file containing the output of Malt Parser (see \url{http://alpage.inria.fr/statgram/frdep/fr_stat_dep_malt.html})}
}
\value{
A list of data frames which can be analyzed using \code{getUnits()} or \code{getMeasures()}.
}
\description{
Extract data frames from output of Malt Parser
}
