% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synt.units.fcn.R
\name{synt.units.fcn}
\alias{synt.units.fcn}
\title{Generate a list of syntactic units}
\usage{
synt.units.fcn(sentence)
}
\arguments{
\item{sentence}{A data frame which contains a sentence in CONLL format}
}
\description{
Generate a list of syntactic units
}
